backend/requirements.txt
# Web
fastapi
uvicorn

# Database
sqlalchemy
psycopg2-binary
alembic

# Auth
supabase
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Utils
pydantic
pydantic[email]
python-dotenv
requests

sqlalchemy-utils